requirements.txt
# UI and Web Interface
streamlit

# Vector Database and Retrieval
faiss-cpu
langchain
langchain-huggingface
langchain-community
langchain-core

# Embeddings and NLP
sentence-transformers
transformers
numpy

# Utilities
python-dotenv
requests
urllib3